https://github.com/mP1/walkingkooka-spreadsheet/pull/8368 net package extracted
- https://github.com/mP1/walkingkooka-spreadsheet/pull/8368
- net package extracted
